[red-knot] Use memory address as AST node key
During type inference on some invalid syntax examples, we encounter situations
where the exact same expression appears in multiple places inside the AST. The
smallest example I could find is this:
```py
for
```

When parsing it, we get the following AST. Note that the exact same `Name(…)`
expression appears both as `target` and as `iter` field on `StmtFor`, once with
`Store` context, once with `Invalid` context:
```
Module(
    ModModule {
        range: 0..3,
        body: [
            For(
                StmtFor {
                    range: 0..3,
                    is_async: false,
                    target: Name(
                        ExprName {
                            range: 3..3,
                            id: Name(""),
                            ctx: Store,
                        },
                    ),
                    iter: Name(
                        ExprName {
                            range: 3..3,
                            id: Name(""),
                            ctx: Invalid,
                        },
                    ),
                    body: [],
                    orelse: [],
                },
            ),
        ],
    },
)
```

This leads to a subsequent failing double-store assertion in
`store_expression_type` because we infer types for both the `target` and the
`iter` expressions.

Here, we fix this by using the memory address of the nodes instead of
relying on source range and kind.