[red-knot] Semantic index: handle invalid `break`s
This fix intends to address panics related to invalid syntax like the
following where a `break` statement is used in a nested definition
inside a loop:

```py
while True:

    def b():
        x: int

        break
```

Note: I don't like the fact that we add even more mutable state to
`SemanticIndexBuilder`. There might very well be a better solution here.
And maybe we need something more general to handle other cases of
invalid syntax. I still thought it would be valuable to propose this
change with relevant regression tests, so we can start thinking about
what would be required to properly fix this.

closes #14342