[red-knot] Add `typing.Any` as a spelling for the Any type
We already had a representation for the Any type, which we would use
e.g. for expressions without type annotations.  We now recognize
`typing.Any` as a way to refer to this type explicitly.  Like other
special forms, this is tracked correctly through aliasing, and isn't
confused with local definitions that happen to have the same name.