[red-knot] Preliminary support for TOML configs in MD tests
This adds very limited support for specifying the target Python version
from a Markdown test. It is an ad-hoc solution, but designed to be
future-compatible. For now, you can add a single TOML code section
anywhere in a Markdown test. It has the following format:

````markdown
```toml
[tool.knot.environment]
target-version = "3.13"
```
````

So far, there is nothing else that can be configured, but it should be
straightforward to extend this to things like a custom typeshed path.